+++ docs/components/Layout/PageLayout.tsx
@@ -1,8 +1,8 @@
 import { responsiveVal } from '@graphcommerce/next-ui'
 import { makeStyles, Theme } from '@material-ui/core'
 import React from 'react'
+import SidebarMenu from '../../components/SidebarMenu'
 import { DirectoryTree } from '../../util/files'
-import SidebarMenu from '../SidebarMenu'
 
 export type LayoutProps = { menuData: DirectoryTree; children: React.ReactNode }
 
@@ -10,24 +10,19 @@ const useStyles = makeStyles(
   (theme: Theme) => ({
     grid: {
       display: 'grid',
-      gridTemplateColumns: '1fr 4fr',
-      gridColumnGap: theme.spacings.md,
+      gridTemplateColumns: `${responsiveVal(150, 300)} 4fr`,
       maxWidth: '100%',
+      overflow: 'hidden',
     },
     side: {
-      position: 'sticky',
-      top: theme.appShell.appBarHeightMd,
-      height: '100vh',
-      '@supports (-webkit-touch-callout: none)': {
-        height: '-webkit-fill-available',
-      },
-      minWidth: responsiveVal(150, 300),
+      overflowY: 'auto',
+      minWidth: 'min-content',
       borderRight: `1px solid ${theme.palette.divider}`,
     },
     content: {
-      '& > *': {
-        wordBreak: 'break-all',
-      },
+      overflow: 'hidden',
+      maxWidth: '100%',
+      padding: `0 ${theme.spacings.sm} ${theme.spacings.sm}`,
     },
     listItemText: {
       ...theme.typography.caption,
@@ -38,7 +33,7 @@ const useStyles = makeStyles(
   },
 )
 
-export default function PageLayout(props: LayoutProps) {
+export default function Layout(props: LayoutProps) {
   const { menuData, children } = props
   const classes = useStyles()