+++ docs/pages/read/[...url].tsx
@@ -1,21 +1,39 @@
-import fs from 'fs'
 import { PageOptions } from '@graphcommerce/framer-next-pages'
+import { Link } from '@material-ui/core'
+import fs from 'fs'
 import { MDXRemote, MDXRemoteSerializeResult } from 'next-mdx-remote'
 import { serialize } from 'next-mdx-remote/serialize'
 import React from 'react'
+import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter'
 import { LayoutFull, LayoutFullProps } from '../../components/Layout/LayoutFull'
 import PageLayout, { LayoutProps } from '../../components/Layout/PageLayout'
-import sanitizeDirectoryTree from '../../components/SidebarMenu/sanitizeDirectoryTree'
+import MDXWrapper from '../../components/MDXWrapper'
+import NextPrevButtons from '../../components/NextPrevButtons'
+import { sanitizeDirectoryTree } from '../../components/SidebarMenu/sanitizeDirectoryTree'
 import { getAbsoluteFilePath, getDirectoryTree } from '../../util/files'
 
-type PageProps = LayoutProps & { compiledMdxSource: MDXRemoteSerializeResult; title: string }
+type PageProps = LayoutProps & { compiledMdxSource: MDXRemoteSerializeResult }
 
 function ArticlePage(props: PageProps) {
-  const { menuData, compiledMdxSource, title } = props
+  const { menuData, compiledMdxSource } = props
+
+  const components = {
+    a: Link,
+    pre: ({ children }) => {
+      return (
+        <SyntaxHighlighter language={children.props.className?.split('-')[1] ?? 'tsx'}>
+          {children.props.children}
+        </SyntaxHighlighter>
+      )
+    },
+  }
 
   return (
     <PageLayout menuData={menuData}>
-      <MDXRemote {...compiledMdxSource} />
+      <MDXWrapper>
+        <MDXRemote {...compiledMdxSource} components={components} />
+      </MDXWrapper>
+      <NextPrevButtons menuData={menuData} />
     </PageLayout>
   )
 }
@@ -27,21 +45,24 @@ ArticlePage.pageOptions = pageOptions
 
 export default ArticlePage
 
-export const getStaticPaths = () =>
+export const getStaticPaths = () => {
   // todo
-  ({
-    paths: [{ params: { url: ['/'] } }],
+  return {
+    paths: [{ params: { url: ['/1-1/getting-started/intro'] } }],
     fallback: 'blocking',
-  })
+  }
+}
 
 export const getStaticProps = async ({ params }) => {
   const { url } = params
   const documentationTree = getDirectoryTree('content')
 
-  if (url.length !== 2) return { notFound: true }
+  if (url.length !== 3) return { notFound: true }
 
-  const sectionDir = url[0]
-  const articleName = url[1]
+  const [chapter, paragraph] = url[0].split('-')
+  const sectionDir = `${chapter}-${url[1]}`
+  const articleName = `${paragraph}-${url[2]}`
+
   const mdxPath = getAbsoluteFilePath(`content/${sectionDir}/${articleName}.mdx`)
 
   if (fs.existsSync(mdxPath)) {
@@ -49,7 +70,6 @@ export const getStaticProps = async ({ params }) => {
       props: {
         menuData: sanitizeDirectoryTree(documentationTree),
         compiledMdxSource: await serialize(fs.readFileSync(mdxPath).toString()),
-        title: url[1],
       },
     }
   }