+++ packages/magento-product/components/ProductSidebarDelivery/index.tsx
@@ -1,5 +1,5 @@
-import { SvgImage, responsiveVal, iconBox, SvgImageSimple } from '@graphcommerce/next-ui'
-import { makeStyles, Theme, Typography } from '@material-ui/core'
+import { SvgImage, responsiveVal, iconOrderBefore, SvgImageSimple } from '@graphcommerce/next-ui'
+import { darken, lighten, makeStyles, Theme, Typography } from '@material-ui/core'
 import React from 'react'
 
 const useStyles = makeStyles(
@@ -14,16 +14,16 @@ const useStyles = makeStyles(
       gridTemplateColumns: `min-content auto`,
       columnGap: theme.spacings.xxs,
       marginTop: theme.spacings.xxs,
-      background: theme.palette.background.paper,
-      border: `1px solid ${theme.palette.divider}`,
+      background:
+        theme.palette.type === 'light'
+          ? darken(theme.palette.background.default, 0.01)
+          : lighten(theme.palette.background.default, 0.2),
       padding: theme.spacings.xxs,
-      borderRadius: responsiveVal(4, 6),
+      borderRadius: responsiveVal(theme.shape.borderRadius * 3, theme.shape.borderRadius * 4),
     },
     text: {},
     image: {
       gridArea: 'image',
-      width: responsiveVal(18, 24),
-      height: responsiveVal(18, 24),
     },
     title: {
       gridArea: 'title',
@@ -42,7 +42,7 @@ export default function ProductSidebarDelivery() {
 
   return (
     <div className={classes.root}>
-      <SvgImageSimple className={classes.image} src={iconBox} />
+      <SvgImageSimple className={classes.image} src={iconOrderBefore} size='small' />
       <Typography className={classes.title} variant='body2' component='div'>
         Order before 22:00
       </Typography>