+++ packages/next-ui/Row/HeroBanner/index.tsx
@@ -1,4 +1,4 @@
-import { ContainerProps, Theme, makeStyles } from '@material-ui/core'
+import { ContainerProps, Theme, makeStyles, useTheme, useMediaQuery } from '@material-ui/core'
 import { m, useTransform, useViewportScroll } from 'framer-motion'
 import React from 'react'
 import Row from '..'
@@ -7,10 +7,6 @@ import { responsiveVal } from '../../Styles/responsiveVal'
 
 const useStyles = makeStyles(
   (theme: Theme) => ({
-    container: {
-      paddingLeft: 0,
-      paddingRight: 0,
-    },
     wrapper: {
       position: 'relative',
     },
@@ -22,21 +18,19 @@ const useStyles = makeStyles(
       justifyItems: 'center',
       alignContent: 'center',
       padding: `${theme.spacings.lg} ${theme.spacings.md}`,
+      paddingTop: `calc(${theme.spacings.lg} - ${theme.spacings.md})`,
       minHeight: `calc(100vh - ${theme.appShell.headerHeightSm})`,
-      '@supports (-webkit-touch-callout: none)': {
-        minHeight: '-webkit-fill-available',
-      },
       '& > *': {
         zIndex: 1,
         maxWidth: 'max-content',
       },
-
       [theme.breakpoints.up('md')]: {
         width: '70%',
         minHeight: `calc(100vh - ${theme.appShell.headerHeightMd})`,
       },
       [theme.breakpoints.up('lg')]: {
         padding: `${theme.spacings.lg} ${theme.spacings.lg}`,
+        paddingTop: `calc(${theme.spacings.lg} - ${theme.spacings.md})`,
         width: '50%',
       },
     },
@@ -54,10 +48,17 @@ const useStyles = makeStyles(
         objectFit: 'cover',
         width: '100%',
         height: '100%',
+        [theme.breakpoints.down('sm')]: {
+          borderRadius: responsiveVal(theme.shape.borderRadius * 2, theme.shape.borderRadius * 3),
+        },
       },
       [theme.breakpoints.up('md')]: {
         height: '100%',
       },
+    },
+    animated: {
+      borderRadius: responsiveVal(theme.shape.borderRadius * 2, theme.shape.borderRadius * 3),
+      overflow: 'hidden',
     },
   }),
   { name: 'HeroBanner' },
@@ -73,23 +74,29 @@ export type HeroBannerProps = UseStyles<typeof useStyles> &
 export default function HeroBanner(props: HeroBannerProps) {
   const { pageLinks, videoSrc, children, ...containerProps } = props
   const classes = useStyles(props)
-
+  const theme = useTheme()
   const { scrollY } = useViewportScroll()
-  const actionsAnimWidth = useTransform(
+  const width = useTransform(
     scrollY,
     [10, 150],
-    [`calc(100% - ${responsiveVal(30, 60)})`, `calc(100% - ${responsiveVal(0, 0)})`],
+    [`calc(100% - ${responsiveVal(20, 60)}))`, `calc(100% - ${responsiveVal(0, 0)})`],
+  )
+  const matches = useMediaQuery(theme.breakpoints.down('sm'))
+  const borderRadius = useTransform(
+    scrollY,
+    [10, 150],
+    [`${responsiveVal(8, 12)}`, `${responsiveVal(0, 0)}`],
   )
 
   return (
-    <Row maxWidth={false} {...containerProps} className={classes.container}>
+    <Row maxWidth={false} {...containerProps} disableGutters>
       <div className={classes.wrapper}>
         <div className={classes.copy}>
           {children}
           {pageLinks}
         </div>
         <div className={classes.asset}>
-          <m.div style={{ width: actionsAnimWidth }}>
+          <m.div style={{ width: !matches ? width : 0, borderRadius }} className={classes.animated}>
             <video src={videoSrc} autoPlay muted loop playsInline disableRemotePlayback />
           </m.div>
         </div>